Removing IDFromName Functions
These functions will be moved to https://github.com/gophercloud/utils.
