+++ acceptance/openstack/networking/v2/networking.go
@@ -528,3 +528,41 @@ func WaitForPortToCreate(client *gophercloud.ServiceClient, portID string, secs 
 		return false, nil
 	})
 }
+
+// This is duplicated from https://github.com/gophercloud/utils
+// so that Gophercloud "core" doesn't have a dependency on the
+// complementary utils repository.
+func IDFromName(client *gophercloud.ServiceClient, name string) (string, error) {
+	count := 0
+	id := ""
+
+	listOpts := networks.ListOpts{
+		Name: name,
+	}
+
+	pages, err := networks.List(client, listOpts).AllPages()
+	if err != nil {
+		return "", err
+	}
+
+	all, err := networks.ExtractNetworks(pages)
+	if err != nil {
+		return "", err
+	}
+
+	for _, s := range all {
+		if s.Name == name {
+			count++
+			id = s.ID
+		}
+	}
+
+	switch count {
+	case 0:
+		return "", gophercloud.ErrResourceNotFound{Name: name, ResourceType: "network"}
+	case 1:
+		return id, nil
+	default:
+		return "", gophercloud.ErrMultipleResourcesFound{Name: name, Count: count, ResourceType: "network"}
+	}
+}
