+++ openstack/compute/v2/servers/testing/requests_test.go
@@ -203,22 +203,6 @@ func TestCreateServerWithUserdataEncoded(t *testing.T) {
 	th.CheckDeepEquals(t, ServerDerp, *actual)
 }
 
-func TestCreateServerWithImageNameAndFlavorName(t *testing.T) {
-	th.SetupHTTP()
-	defer th.TeardownHTTP()
-	HandleServerCreationSuccessfully(t, SingleServerBody)
-
-	actual, err := servers.Create(client.ServiceClient(), servers.CreateOpts{
-		Name:          "derp",
-		ImageName:     "cirros-0.3.2-x86_64-disk",
-		FlavorName:    "m1.tiny",
-		ServiceClient: client.ServiceClient(),
-	}).Extract()
-	th.AssertNoErr(t, err)
-
-	th.CheckDeepEquals(t, ServerDerp, *actual)
-}
-
 func TestDeleteServer(t *testing.T) {
 	th.SetupHTTP()
 	defer th.TeardownHTTP()
@@ -358,7 +342,7 @@ func TestRebuildServer(t *testing.T) {
 	opts := servers.RebuildOpts{
 		Name:       "new-name",
 		AdminPass:  "swordfish",
-		ImageID:    "http://104.130.131.164:8774/fcad67a6189847c4aecfa3c81a05783b/images/f90f6034-2570-4974-8351-6b49732ef2eb",
+		ImageRef:   "f90f6034-2570-4974-8351-6b49732ef2eb",
 		AccessIPv4: "1.2.3.4",
 	}
